+++ src/sdr/sdr.cpp
@@ -3,7 +3,7 @@
 #include "sdr.h"
 #include "../constants.h"
 
-void SDR::poll(std::complex<float> *samples) {
+void SDR::poll(std::complex<double> *samples) {
     this->buffer.pop(samples, FFT_WINDOW_SIZE);
 }
